refactor: rename Firo RPC references to network-agnostic names, fix file casing
