+++ .changeset/light-pens-beam.md
@@ -1,0 +1,5 @@
+---
+'@rosen-bridge/firo-scanner': minor
+---
+
+Add FiroElectrumXNetwork TCP connector and FiroElectrumXScanner, replace firod JSON-RPC dependency with ElectrumX protocol
