+++ .changeset/quiet-masks-reply.md
@@ -1,0 +1,5 @@
+---
+'@rosen-bridge/firo-observation-extractor': minor
+---
+
+Rename FiroRpcObservationExtractor to FiroObservationExtractor, deprecate RPC-specific alias
