+++ packages/observation-extractors/firo-observation-extractor/lib/index.ts
@@ -1,1 +1,1 @@
-export * from './firoRpcObservationExtractor';
+export * from './firoObservationExtractor';
