+++ packages/scanners/firo-scanner/lib/index.ts
@@ -1,3 +1,3 @@
-export { FiroElectrumXScanner } from './scanner/firoElectrumxScanner';
-export { FiroElectrumXNetwork } from './network/firoElectrumxNetwork';
+export { FiroElectrumXScanner } from './scanner/firoElectrumXScanner';
+export { FiroElectrumXNetwork } from './network/firoElectrumXNetwork';
 export * from './types';
